+++ tests/GPUSystemTest/src/GPUSystemTestProbe.cpp
@@ -35,11 +35,11 @@ int GPUSystemTestProbe::outputState(double timed){
    const pvdata_t * A = getTargetLayer()->getLayerData();
    float sumsq = 0;
    for (int i = 0; i < numExtNeurons; i++){
-      pvErrorIf(!(fabs(A[i]) < 5e-4), "Test failed.\n");
+      pvErrorIf(!(fabsf(A[i]) < 5e-4f), "Test failed.\n");
    }
    for(int b = 0; b < loc->nbatch; b++){
       //For max std of 5e-5
-      pvErrorIf(!(sigma[b] <= 5e-5), "Test failed.\n");
+      pvErrorIf(!(sigma[b] <= 5e-5f), "Test failed.\n");
    }
 
    return status;
