+++ tests/GroupNormalizationTest/src/AllConstantValueProbe.cpp
@@ -40,8 +40,8 @@ int AllConstantValueProbe::outputState(double timed) {
    if (this->parent->columnId()==0) {
       for(int b = 0; b < this->parent->getNBatch(); b++){
          if (timed>0 && (fMin[b]<correctValue-nnzThreshold || fMax[b] > correctValue+nnzThreshold)) {
-            outputStream->printf("     Values outside of tolerance nnzThreshold=%f\n", nnzThreshold);
-            pvErrorNoExit().printf("t=%f: fMin=%f, fMax=%f; values more than nnzThreshold=%g away from correct value %f\n", timed, fMin[b], fMax[b], nnzThreshold, correctValue);
+            outputStream->printf("     Values outside of tolerance nnzThreshold=%f\n", (double)nnzThreshold);
+            pvErrorNoExit().printf("t=%f: fMin=%f, fMax=%f; values more than nnzThreshold=%g away from correct value %f\n", timed, (double)fMin[b], (double)fMax[b], (double)nnzThreshold, (double)correctValue);
          }
       }
    }
