+++ tests/ReceiveFromPostTest/src/ReceiveFromPostProbe.cpp
@@ -18,7 +18,7 @@ ReceiveFromPostProbe::ReceiveFromPostProbe(const char * probeName, HyPerCol * hc
 }
 
 int ReceiveFromPostProbe::initReceiveFromPostProbe_base() {
-   tolerance = (pvadata_t) 1e-3;
+   tolerance = (pvadata_t) 1e-3f;
    return PV_SUCCESS;
 }
 
@@ -46,16 +46,16 @@ int ReceiveFromPostProbe::outputState(double timed){
    int numExtNeurons = getTargetLayer()->getNumExtended();
    const pvdata_t * A = getTargetLayer()->getLayerData();
    for (int i = 0; i < numExtNeurons; i++){
-      if(fabs(A[i]) != 0){
+      if(fabsf(A[i]) != 0){
          int xpos = kxPos(i, loc->nx+loc->halo.lt+loc->halo.rt, loc->ny+loc->halo.dn+loc->halo.up, loc->nf);
          int ypos = kyPos(i, loc->nx+loc->halo.lt+loc->halo.rt, loc->ny+loc->halo.dn+loc->halo.up, loc->nf);
          int fpos = featureIndex(i, loc->nx+loc->halo.lt+loc->halo.rt, loc->ny+loc->halo.dn+loc->halo.up, loc->nf);
          //pvInfo() << "[" << xpos << "," << ypos << "," << fpos << "] = " << std::fixed << A[i] << "\n";
       }
       //For roundoff errors
-      if(fabs(A[i]) >= tolerance) {
+      if(fabsf(A[i]) >= tolerance) {
          pvErrorNoExit().printf("%s %s activity outside of tolerance %f: extended index %d has activity %f\n",
-               getMessage(), getTargetLayer()->getDescription_c(), tolerance, i, A[i]);
+               getMessage(), getTargetLayer()->getDescription_c(), (double)tolerance, i, (double)A[i]);
          status = PV_FAILURE;
       }
       if (status != PV_SUCCESS) {
