+++ tests/test_kypos/src/test_kypos.cpp
@@ -26,7 +26,7 @@ int main(int argc, char* argv[])
       float ky = kyPos(kl, nxLocal, nyLocal, nf);
 
       if (ky != (float)kk) {
-         pvError().printf("FAILED:TEST_KYPOS: (k,ky) = (%d,%f)\n", kl, ky);
+         pvError().printf("FAILED:TEST_KYPOS: (k,ky) = (%d,%f)\n", kl, (double)ky);
       }
    }
 
@@ -47,7 +47,7 @@ int main(int argc, char* argv[])
       float ky = kyPos(kl, nxLocal, nyLocal, nf);
 
       if ((int)ky-kk != 0) {
-         pvError().printf("FAILED:TEST_KYPOS: (k,ky) = (%d,%f)\n", kl, ky);
+         pvError().printf("FAILED:TEST_KYPOS: (k,ky) = (%d,%f)\n", kl, (double)ky);
       }
    }
 
@@ -69,7 +69,7 @@ int main(int argc, char* argv[])
       float ky = kyPos(kl, nxLocal, nyLocal, nf);
 
       if ((int)ky-kk != 0) {
-         pvError().printf("FAILED:TEST_KYPOS: (k,ky) = (%d,%f)\n", kl, ky);
+         pvError().printf("FAILED:TEST_KYPOS: (k,ky) = (%d,%f)\n", kl, (double)ky);
       }
    }
 
@@ -92,7 +92,7 @@ int main(int argc, char* argv[])
       float ky = kyPos(kl, nxLocal, nyLocal, nf);
 
       if ((int)ky-kk != 0) {
-         pvError().printf("FAILED:TEST_KYPOS: (k,ky) = (%d,%f)\n", kl, ky);
+         pvError().printf("FAILED:TEST_KYPOS: (k,ky) = (%d,%f)\n", kl, (double)ky);
       }
    }
